# cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "✨Reverse-engineered API for Gemini web app. It can be used as a genuine API key from OpenAI, Gemini, and Claude.",
        "title": "Gemini Web To API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/openai/v1/chat/completions": {
            "post": {
                "description": "Generates a completion for the chat message",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAI"
                ],
                "summary": "Chat Completions (OpenAI)",
                "parameters": [
                    {
                        "description": "Chat Completion Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/openai/v1/images/generations": {
            "post": {
                "description": "Creates an image given a prompt",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAI"
                ],
                "summary": "Image Generations (OpenAI)",
                "parameters": [
                    {
                        "description": "Image Generation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ImageGenerationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ImageGenerationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/openai/v1/responses": {
            "post": {
                "description": "Alias for chat completions for specific external integration requirements",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "OpenAI"
                ],
                "summary": "Chat Responses (OpenAI Alias)",
                "parameters": [
                    {
                        "description": "Chat Completion Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ChatCompletionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ChatCompletionRequest": {
            "type": "object",
            "properties": {
                "max_tokens": {
                    "type": "integer"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                },
                "model": {
                    "type": "string"
                },
                "response_format": {
                    "$ref": "#/definitions/dto.ResponseFormat"
                },
                "stream": {
                    "type": "boolean"
                },
                "temperature": {
                    "type": "number"
                },
                "tools": {
                    "type": "array",
                    "items": {}
                }
            }
        },
        "dto.ChatCompletionResponse": {
            "type": "object",
            "properties": {
                "choices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.Choice"
                    }
                },
                "created": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "object": {
                    "type": "string"
                },
                "system_fingerprint": {},
                "usage": {
                    "$ref": "#/definitions/models.Usage"
                }
            }
        },
        "dto.Choice": {
            "type": "object",
            "properties": {
                "finish_reason": {
                    "type": "string"
                },
                "index": {
                    "type": "integer"
                },
                "message": {
                    "$ref": "#/definitions/models.Message"
                }
            }
        },
...
        "dto.ImageData": {
            "type": "object",
            "properties": {
                "b64_json": {
                    "type": "string"
                },
                "revised_prompt": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "dto.ImageGenerationRequest": {
            "type": "object",
            "properties": {
                "model": {
                    "type": "string"
                },
                "n": {
                    "type": "integer"
                },
                "prompt": {
                    "type": "string"
                },
                "quality": {
                    "type": "string"
                },
                "response_format": {
                    "description": "\"url\" or \"b64_json\"",
                    "type": "string"
                },
                "size": {
                    "type": "string"
                },
                "style": {
                    "type": "string"
                },
                "user": {
                    "type": "string"
                }
            }
        },
        "dto.ImageGenerationResponse": {
            "type": "object",
            "properties": {
                "created": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ImageData"
                    }
                }
            }
        },
...
        "dto.ResponseFormat": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.FunctionCall": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "tool_call_id": {
                    "type": "string"
                },
                "tool_calls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ToolCall"
                    }
                }
            }
        },
...
        "models.ToolCall": {
            "type": "object",
            "properties": {
                "function": {
                    "$ref": "#/definitions/models.FunctionCall"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.Usage": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "completion_tokens_details": {},
                "input_tokens": {
                    "type": "integer"
                },
                "output_tokens": {
                    "type": "integer"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "prompt_tokens_details": {},
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \u003cyour_admin_api_key\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}